s/_header/_headerSent/g
The _header field can be set in a few ways that don't necessarily
mean that it's been sent.  Also, for HTTP 0.9 requests, it should
be set to an empty string, even when set (this is currently not
handled correctly in node, but will be eventually.)

Replace checks for _headerSent, which is a boolean indicating that
the header portion of the request has actually been sent to the
client, which is the intended semantics here.